refactoring sign methods
